added min and mean added N input support for min/max/mean
