+++ onnx2keras/layers.py
@@ -1,8 +1,8 @@
 from .convolution_layers import convert_conv, convert_convtranspose
 from .activation_layers import convert_relu, convert_lrelu, convert_selu, \
     convert_sigmoid, convert_tanh, convert_softmax
-from .operation_layers import convert_clip, convert_exp, convert_reduce_sum, convert_reduce_mean, convert_log, convert_pow, convert_sqrt, convert_split, convert_max
-from .elementwise_layers import convert_elementwise_div, convert_elementwise_add, convert_elementwise_mul, convert_elementwise_sub
+from .operation_layers import convert_clip, convert_exp, convert_reduce_sum, convert_reduce_mean, convert_log, convert_pow, convert_sqrt, convert_split
+from .elementwise_layers import convert_elementwise_div, convert_elementwise_add, convert_elementwise_mul, convert_elementwise_sub, convert_max, convert_min, convert_mean
 from .linear_layers import convert_gemm
 from .reshape_layers import convert_transpose, convert_shape, convert_gather, convert_unsqueeze, convert_concat, convert_reshape, convert_flatten, convert_slice
 from .constant_layers import convert_constant
@@ -29,6 +29,8 @@ AVAILABLE_CONVERTERS = {
     'Sqrt': convert_sqrt,
     'Split': convert_split,
     'Max': convert_max,
+    'Min': convert_min,
+    'Mean': convert_mean,
     'Div': convert_elementwise_div,
     'Add': convert_elementwise_add,
     'Sum': convert_elementwise_add,
